Connect telegram bot with backend (#12)

+++ bot/src/main/scala/com/github/mmvpm/bot/model/Draft.scala
@@ -4,7 +4,7 @@ import com.github.mmvpm.model.OfferDescription
 
 case class Draft(
     name: Option[String] = None,
-    price: Option[Long] = None,
+    price: Option[Int] = None,
     description: Option[String] = None,
     photos: Seq[String] = Seq.empty
 ) {
@@ -14,5 +14,5 @@ case class Draft(
       name <- name
       price <- price
       description <- description
-    } yield OfferDescription(name, price.toInt, description)
+    } yield OfferDescription(name, price, description)
 }
